Override PIDController::Calculate
This change was not as simple as it should have been because the Calculate
method is not virtual, and the variables needed to perform the calculation are
declared private. Therefore, I copy-and-pasted the WPILib class into our code
base and modified it to correct those issues, then overrode our new class.